(hds-2383) Remove old cookie consent files
