--- packages/react/src/components/cookieConsent/index.ts
@@ -1,4 +1,0 @@
-export * from './cookieModal/CookieModal';
-export * from './cookiePage/CookiePage';
-export * from './useCookies';
-export * from './types';
