--- packages/react/src/components/cookieConsent/languageSwitcher/LanguageSwitcherItem/LanguageSwitcherItem.module.scss
@@ -1,157 +1,0 @@
-@import '../../../../styles/common.scss';
-@value x-small-down, small-down, small-only, medium-only, medium-up, large-only, large-up, x-large-only, x-large-up from "../../../../styles/breakpoints.scss";
-
-%item {
-  text-decoration: none;
-}
-
-.item {
-  @extend %item;
-
-  &:focus {
-    transition: none;
-  }
-}
-
-.primary {
-  @extend %item;
-}
-
-.secondary {
-  @extend %item;
-}
-
-.supplementary {
-  @extend %item;
-
-  --border-width: 3px;
-}
-
-.primary:focus,
-.secondary:focus,
-.supplementary:focus {
-  transition: none;
-}
-
-.icon {
-  &#{&} {
-    display: flex;
-    height: unset;
-    width: unset;
-  }
-}
-
-.label {
-  padding: 0;
-}
-
-.rowItem,
-.dropdownItem,
-.dropdownLinkItem {
-  align-items: center;
-  box-sizing: border-box;
-  color: inherit;
-  display: flex;
-  font-weight: 500;
-  outline: none;
-  text-decoration: none;
-}
-
-.rowItem {
-  @extend %buttonReset;
-
-  .icon {
-    margin: 0 var(--spacing-2-xs) 0 calc(var(--spacing-3-xs) * -1);
-  }
-
-  .label {
-    line-height: var(--lineheight-s);
-    margin: 0;
-  }
-}
-
-@media small-down {
-  .rowItem {
-    height: var(--spacing-xl);
-    padding: var(--spacing-2-xs);
-
-    &:focus {
-      box-shadow: 0 0 0 3px var(--color-coat-of-arms);
-    }
-
-    &:not(:first-child) {
-      margin-top: var(--spacing-2-xs);
-    }
-  }
-
-  .dropdownItem,
-  .dropdownLinkItem {
-    &:not(:first-child) {
-      margin-top: var(--spacing-2-xs);
-    }
-  }
-
-  .icon.icon {
-    align-self: center;
-  }
-}
-
-@media medium-up {
-  .rowItem {
-    padding: var(--spacing-3-xs);
-    position: relative;
-  }
-
-  // focus outline
-  .rowItem:focus:before,
-  .dropdownItem:focus-within:before,
-  .dropdownLinkItem a:focus-within:before,
-  .dropdownLinkItem button:focus-within:before {
-    border-radius: 0;
-    box-shadow: 0 0 0 3px var(--color-coat-of-arms);
-    content: '';
-    height: calc(100% - 3px * 2);
-    left: 0;
-    pointer-events: none;
-    position: absolute;
-    top: 0;
-    transform: translateY(calc(3px));
-    width: 100%;
-    z-index: 1;
-  }
-
-  .dropdownLinkItem a,
-  .dropdownLinkItem button {
-    position: relative;
-  }
-
-  // focus outline - active
-  .rowItem:active:before,
-  .rowItem.active:before,
-  .dropdownItem.active:before,
-  .dropdownLinkItem.active:before {
-    height: calc(100% - 3px);
-  }
-
-  // active + hover styles
-  .rowItem:active:after,
-  .rowItem:hover:not(:focus):after,
-  .rowItem.active:after,
-  .dropdownItem:hover:not(:focus-within):after,
-  .dropdownItem.active:after,
-  .dropdownLinkItem:hover:not(:focus-within):after,
-  .dropdownLinkItem.active:after {
-    background-color: var(--color-black-90);
-    bottom: 0;
-    content: '';
-    height: var(--spacing-3-xs);
-    left: 0;
-    pointer-events: none;
-    position: absolute;
-    width: 100%;
-  }
-
-  .dropdownItem:not(.icon) {
-    width: max-content;
-  }
-}
